lab8: ghi chu realtime - Vue 3 + WebSocket + SQL Server NotesDB
De chinh (7 muc) va de phu (5 API) deu hoan chinh.

Backend (demob1):
- Xoa package LAB/ cua lab7 (chat) gom 14 class, trong do co WebSocketConfig
  thu hai bi trung @EnableWebSocketMessageBroker; xoa dto/, entity/,
  templates/index.html va dependency thymeleaf khong con dung
- Cau truc lai theo model/repository/service/controller,
  NoteService tach interface + Impl
- Doi tu H2 sang SQL Server, dung dung bang Notes cua NotesDB.sql:
  id INT IDENTITY -> Integer, NVARCHAR -> @Nationalized, ddl-auto=none
  nen Hibernate khong tu tao hay sua schema
- NoteController bind thang @RequestBody Note, bo 2 record DTO
- Validate author/content rong -> 400, khong tim thay -> 404

Frontend (l8-fe):
- Them sockjs-client va @stomp/stompjs con thieu trong package.json
- Them define global -> window trong vite.config.js cho sockjs-client
- Bo @keyup.enter trung voi form submit (Enter dang gui ghi chu 2 lan)
- Them co sending chong double-click va thong bao loi khi mat backend

Kiem thu: mvnw test 9 tests pass - 5 API cua de phu, truong hop 404/400,
thu tu sap xep giam dan, va NoteWebSocketBroadcastTests noi STOMP that vao
/ws de kiem chung yeu cau 4. Test chay tren H2 MODE=MSSQLServer voi dung
schema NotesDB.sql nen khong can bat SQL Server. npm run build sach.
Chua chay thu tren SQL Server that (may viet code khong cai SQL Server).

Huong dan chay, chuan bi NotesDB va kiem thu o manual_lab8.txt.


Bo sung:
- CORS /api/** mo cho MOI origin (truoc do chi cho http://localhost:*),
  co test preflight tu 3 origin khac nhau
- Frontend: bat loi khi tai danh sach that bai, chan hien trung theo id,
  tu them ghi chu khi socket dang rot va tai lai danh sach khi noi lai,
  hien canh bao khi mat ket noi realtime

- Doi title trang tu "Vite App" thanh "LAB 8 - Ghi chu realtime", lang=vi

+++ demob1/src/main/java/com/vn/test/demob1/repository/NoteRepository.java
@@ -1,12 +1,12 @@
 package com.vn.test.demob1.repository;
 
-import com.vn.test.demob1.entity.Note;
+import com.vn.test.demob1.model.Note;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface NoteRepository extends JpaRepository<Note, Long> {
+public interface NoteRepository extends JpaRepository<Note, Integer> {
 
-    // req 1 (đề phụ): danh sách ghi chú, sắp xếp theo thời gian tạo giảm dần
+    // đề phụ #1: danh sách ghi chú sắp xếp theo thời gian tạo giảm dần
     List<Note> findAllByOrderByCreatedAtDesc();
 }
